+++ plugins/amoeba/serialization/src/AmoebaStretchBendForceProxy.cpp
@@ -45,7 +45,7 @@ void AmoebaStretchBendForceProxy::serialize(const void* object, SerializationNod
     node.setIntProperty("version", 1);
     const AmoebaStretchBendForce& force = *reinterpret_cast<const AmoebaStretchBendForce*>(object);
     SerializationNode& bonds = node.createChildNode("StretchBendAngles").setIntProperty( "size", force.getNumStretchBends() );
-    for (unsigned int ii = 0; ii < force.getNumStretchBends(); ii++) {
+    for (unsigned int ii = 0; ii < static_cast<unsigned int>(force.getNumStretchBends()); ii++) {
         int particle1, particle2, particle3;
         double distanceAB, distanceCB, angle, k;
         force.getStretchBendParameters(ii, particle1, particle2, particle3, distanceAB, distanceCB, angle, k);
